Uploaded .ipynb containing the required code.
deepface.ipynb : Contains code to run and test deepface on various backend detectors , as well as create profiles for known members using their images. Allows to save a .npy file for each person's embedding storing their 'profile'.
Domestic_Theft_Prevention.ipynb : Stores the main logic to run the surveillance system on a video feed ( can be extended to run on a live feed using multiple servers). Simulates the action of opening the door by using compass sensor data( if change in degrees from initial configuration is above a certain threshold, it is classified as motion of the attached camera(on the door) and hence an attempt to open the gate. Once such an action is done, face recognition on previous few seconds is run, if any known person is NOT found, an alert is sent .